fix(runtime): remove unused upstream proxy helper
